# Conflicts:
#	sources/kernel/prototypes.cl

+++ sources/kernel/pathtrace_algo/pathtrace.cl
@@ -12,7 +12,7 @@ float3		pathtrace(
 		__global const float3 *v_textures,
 		__global const t_renderer_params *params,
 		__global const t_texture_info *texture_info,
-		__global const float *texture_list,
+		__global const int *texture_list,
 		t_ray ray,
 		int depth,
 		float *seed,
